파일 업로드, 다운로드 완료
excel 추출 구현 중 저장

+++ nextobe/src/main/java/com/kimyongju/nextobe/board/vo/FileVO.java
@@ -8,7 +8,7 @@ public class FileVO {
 	private int bno;
 	private String org_file_name;
 	private String save_file_name;
-	private int file_size;
+	private Long file_size;
 	
 	
 	/* VAM_FILE
